JPERF-273: Remove fluent hook API
