+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/flow/PreInstallFlow.kt
@@ -1,9 +1,29 @@
 package com.atlassian.performance.tools.infrastructure.api.jira.flow
 
 import com.atlassian.performance.tools.infrastructure.api.jira.flow.server.PreInstallHook
+import com.atlassian.performance.tools.ssh.api.SshConnection
+import java.util.*
+import java.util.concurrent.ConcurrentLinkedQueue
 
-interface PreInstallFlow : PostInstallFlow {
+open class PreInstallFlow protected constructor() : PostInstallFlow() {
+
+    private val preInstallHooks: Queue<PreInstallHook> = ConcurrentLinkedQueue()
+
     fun hook(
         hook: PreInstallHook
-    ): PreInstallFlow
+    ) {
+        preInstallHooks.add(hook)
+    }
+
+    internal fun runPreInstallHooks(
+        ssh: SshConnection,
+        server: TcpServer
+    ) {
+        while (true) {
+            preInstallHooks
+                .poll()
+                ?.run(ssh, server, this)
+                ?: break
+        }
+    }
 }
